Fixed bug: CourseUser course list not saved in database
Former-commit-id: 36ddc3f516a95bff8fd55453c730fcf781702078

+++ src/in/co/praveenkumar/mdroid/moodlemodel/MoodleUserCourse.java
@@ -23,7 +23,7 @@ public class MoodleUserCourse extends SugarRecord<MoodleUserCourse> {
 
 	// Relational fields - for less logout complexity
 	long siteid;
-	long userid;
+	int userid; // Moodle userid
 
 	/**
 	 * Get id of the course
@@ -84,7 +84,7 @@ public class MoodleUserCourse extends SugarRecord<MoodleUserCourse> {
 	 * 
 	 * @return
 	 */
-	public void setUserid(long userid) {
+	public void setUserid(int userid) {
 		this.userid = userid;
 	}
 }
